feat(bridge-react): add en version for bridge document and refactor types
